Sock closes on process exit

+++ user/user.h
@@ -43,5 +43,5 @@ void* malloc(uint);
 void free(void*);
 
 // sock.c
-int sock_open(int);
+int sock_open(uint16);
 int sock_close(int);
